grouping of nodes
