+++ app/routes.py
@@ -7,6 +7,7 @@ import json
 import yaml
 import os
 import threading
+from app.services.group_graph_edge import group_graph
 from app import app, databases, schema_manager, db_instance,socketio
 from app.lib import validate_request
 from flask_cors import CORS
@@ -20,6 +21,7 @@ from app.lib import convert_to_csv
 from app.lib import generate_file_path
 from app.lib import adjust_file_path
 from flask_socketio import send,emit,join_room,leave_room
+
 import json
  
 # Load environmental variables
@@ -261,131 +263,6 @@ def process_query(current_user_id):
     except Exception as e:
         logging.error(f"Error processing query: {e}")
         return jsonify({"error": str(e)}), 500
-def group_graph(result_graph, request):
-    
-    MINIMUM_EDGES_TO_COLLAPSE = 2
-     
-    result_graph=json.loads(result_graph)
-     
-    Edge=result_graph['edges']
-     
-    # Get all unique edge types
-    edge_types = list(set(edge["type"] for edge in request['requests']["predicates"]))
-
- 
-    # Group edges by source or target for each edge type
-    edge_groupings = []
-    for edge_type in edge_types:
-        # Filter edges of this type
-        
-        edges_of_type = [edge for edge  in Edge if edge["data"]["label"] == edge_type]
-        
-        # Group by source and target
-        source_groups = defaultdict(list)
-        target_groups = defaultdict(list)
-        for edge in edges_of_type:
-            source_groups[edge["data"]["source"]].append(edge)
-            target_groups[edge["data"]["target"]].append(edge)
-
-        # Choose grouping with fewer groups
-        grouped_by = "source" if len(source_groups) <= len(target_groups) else "target"
-        groups = source_groups if grouped_by == "source" else target_groups
-         
-        # Save grouping info
-        edge_groupings.append({
-            "count": len(edges_of_type),
-            "edgeType": edge_type,
-            "groupedBy": grouped_by,
-            "groups": groups,
-        })
-
-    # Sort edge groupings to process the most impactful ones first
-    edge_groupings.sort(key=lambda g: g["count"] - len(g["groups"]), reverse=True)
-     
-    # Process each edge grouping to modify the graph
-    new_graph = result_graph.copy()
-    for grouping in edge_groupings:
-        sorted_groups = sorted(grouping["groups"].items(), key=lambda x: len(x[1]), reverse=True)
-        for key, edges in sorted_groups:
-            if len(edges) < MINIMUM_EDGES_TO_COLLAPSE:
-                continue
-             
-            # Get IDs of nodes to be grouped
-            child_node_ids = [
-                edge["data"]["source"] if grouping["groupedBy"] == "target" else edge["data"]["target"]
-                for edge in edges
-            ]
-
-            # Filter child nodes
-            child_nodes = [node for node in new_graph["nodes"] if node["data"]["id"] in child_node_ids]
-            unique_parents = list({node["data"].get("parent") for node in child_nodes if "parent" in node["data"]})
-
-            if len(unique_parents) > 1:
-                continue  # Skip if nodes have different parents
-
-            # Check if the parent has the same child nodes
-            if unique_parents:
-                parent_id = unique_parents[0]
-                all_child_nodes = [
-                    node for node in new_graph["nodes"]
-                    if node["data"].get("parent") == parent_id
-                ]
-                if len(all_child_nodes) == len(child_nodes):
-                    add_new_edge(new_graph, edges[0], parent_id, grouping["groupedBy"])
-                    continue
-             
-            # Create a new parent node
-            parent_id = f"n{uuid.uuid4().hex[:8]}"
-            parent_node = {"data": {"id": parent_id, "type": "parent"}}
-            new_graph["nodes"].append(parent_node)
-
-            # Assign parent to child nodes
-            for node in new_graph["nodes"]:
-                if node["data"]["id"] in child_node_ids:
-                    node["data"]["parent"] = parent_id
-
-            # Add a new edge pointing to the parent node
-            add_new_edge(new_graph, edges[0], parent_id, grouping["groupedBy"])
-             
-    # Step 6: Count types for each parent
-    parent_counts = defaultdict(lambda: defaultdict(int))
-
-    # Iterate through all nodes to check for parent nodes and count based on types
-    for node in new_graph["nodes"]:
-        # Check if the node is a parent node
-        if node["data"].get("type") == "parent":
-            parent_id = node["data"].get("id")  # Get parent ID
-            if parent_id:
-                # Now iterate through all nodes to check the parent and count based on type
-                for child_node in new_graph["nodes"]:
-                    if child_node["data"].get("parent") == parent_id:
-                        node_type = child_node["data"].get("type")
-                        # Increment count based on the node type
-                        if node_type:  # Check if the type exists
-                            parent_counts[parent_id][node_type] += 1
-
-    # Step 7: Update the parent nodes with the counts for each type (promoter, gene, etc.)
-    for parent_id, counts in parent_counts.items():
-        for node in new_graph["nodes"]:
-            if node["data"].get("id") == parent_id:
-                node["data"].update(counts)  # Update the parent node with the counts
-                break
-    return new_graph
-
-
-def add_new_edge(graph, edge, parent_id, grouped_by):
-    new_edge_id = f"e{uuid.uuid4().hex[:8]}"
-    new_edge = {
-        "data": {
-            **edge["data"],
-            "id": new_edge_id,
-            grouped_by: parent_id,
-        }
-    }
-    graph["edges"] = [
-        new_edge
-    ] + [e for e in graph["edges"] if e != edge]
-
